allow single files as lint targets
